fix: align LegacyPacket format_code type with wire format (uint16_t -> uint32_t)
The serialization code uses write_le32/read_le32 for format_code (4 bytes),
but the struct field was uint16_t (2 bytes), causing round-trip data corruption.

+++ mozart_unified/include/mozart/core/contract_packet.hpp
@@ -32,7 +32,7 @@ struct LegacyPacket {
 
     uint32_t seq = 0;
     uint64_t timestamp_us = 0;
-    uint16_t format_code = 0;
+    uint32_t format_code = 0;
     std::vector<int16_t> samples;
 
     std::vector<uint8_t> pack() const;
